exclude blockquotes and comments from count
Option to exclude blockquotes and comments from the count.
Comments are supported both in html and in critic markup format,
and can span one or more lines.
Blockquotes are supported in lazy format, so that any line after
a > until the next empty line is seen as a blockquote. Further,
a > may have zero to three leading spaces.